+++ src/blocks/lcd_screen/virtual-circuit.ts
@@ -10,11 +10,11 @@ import type {
 
 import type { LCDScreenState } from "./state";
 import type { Element, Svg, Text } from "@svgdotjs/svg.js";
-import { positionComponent } from "../../core/virtual-circuit/svg-position-v2";
+import { positionComponent } from "../../core/virtual-circuit/svg-position";
 import {
   createComponentWire,
   createGroundOrPowerWire,
-} from "../../core/virtual-circuit/wire-v2";
+} from "../../core/virtual-circuit/wire";
 
 /**
  * Timer for blinking
