+++ src/blocks/ultrasonic_sensor/virtual-circuit.ts
@@ -11,16 +11,11 @@ import type {
 import type { UltraSonicSensorState } from "./state";
 import type { Element, Svg } from "@svgdotjs/svg.js";
 
-import { positionComponent } from "../../core/virtual-circuit/svg-position-v2";
-import {
-  createWire,
-  createPowerWire,
-  createGroundWire,
-} from "../../core/virtual-circuit/wire";
+import { positionComponent } from "../../core/virtual-circuit/svg-position";
 import {
   createComponentWire,
   createGroundOrPowerWire,
-} from "../../core/virtual-circuit/wire-v2";
+} from "../../core/virtual-circuit/wire";
 
 export const positionUltraSonicSensor: PositionComponent<UltraSonicSensorState> = (
   state,
