+++ src/core/virtual-circuit/svg-position.ts
@@ -1,22 +1,21 @@
 import { findSvgElement } from "./svg-helpers";
 import type { Element, Svg } from "@svgdotjs/svg.js";
-import type { ARDUINO_PINS } from "../microcontroller/selectBoard";
-import type { MicroController } from "../microcontroller/microcontroller";
 
 export const positionComponent = (
   element: Element,
   arduino: Element,
   draw: Svg,
-  wire: ARDUINO_PINS,
-  connectionId: string,
-  board: MicroController
+  hole: number,
+  isDown: boolean,
+  connectionId: string
 ) => {
   // 1 Take the Arduino X position
   // 2 Add to it the hole's x position
   // 3 minus the center of the pin in the virtual component
+  const holeId = `pin${hole}${isDown ? "E" : "F"}`;
   element.x(
     arduino.x() +
-      findSvgElement(board.pinToBreadboardHole(wire), draw).cx() -
+      findSvgElement(holeId, draw).cx() -
       findSvgElement(connectionId, element).cx()
   );
