+++ src/core/virtual-circuit/wire.ts
@@ -1,146 +1,264 @@
-import { findComponentConnection, findSvgElement } from "./svg-helpers";
-import type { Svg, Element, Line } from "@svgdotjs/svg.js";
-import type { ARDUINO_PINS } from "../microcontroller/selectBoard";
+import { Element, Line, Svg } from "@svgdotjs/svg.js";
 import _ from "lodash";
-import type { MicroController } from "../microcontroller/microcontroller";
+import resistorSvg from "../virtual-circuit/commonsvgs/resistors/resistor-small.svg";
 
-export interface Wire {
-  id: string;
-  pointBreadboardId: string;
-  pointComponentId: string;
-}
+import {
+  Breadboard,
+  BreadBoardArea,
+  MicroController,
+} from "../microcontroller/microcontroller";
+import { ANALOG_PINS, ARDUINO_PINS } from "../microcontroller/selectBoard";
+import {
+  findSvgElement,
+  findComponentConnection,
+  findArduinoConnectionCenter,
+} from "./svg-helpers";
 
-export enum WIRE_COLOR {
-  GND = "#000000",
-  POWER = "#FF422A",
-}
+let breadboard: Breadboard = {
+  areas: [],
+  order: [],
+};
 
-export const createWire = (
-  element: Element,
+export const takeBoardArea = (): BreadBoardArea | null => {
+  const areas = breadboard.areas;
+  for (let orderIndex in breadboard.order) {
+    const areaIndex = breadboard.order[orderIndex];
+    const area = areas[areaIndex];
+    if (!area.taken) {
+      breadboard.areas[areaIndex].taken = true;
+      return breadboard.areas[areaIndex];
+    }
+  }
+
+  return null;
+};
+
+export const createWireFromArduinoToBreadBoard = (
   pin: ARDUINO_PINS,
-  connectionId: string,
-  arduino: Element,
+  arduinoEl: Svg,
   draw: Svg,
-  color: string,
-  type: string,
+  breadBoardHoleId,
+  componentId: string,
   board: MicroController
 ) => {
-  const hole = findBreadboardHoleXY(pin, arduino, draw, board);
-  const componentPin = findComponentConnection(element, connectionId);
+  const hole = findBreadboardHoleXY(breadBoardHoleId, arduinoEl, draw);
+  const pinConnection = board.pinConnections[pin];
+  const arduinoPin = findArduinoConnectionCenter(arduinoEl, pinConnection.id);
+
+  const line = draw
+    .line()
+    .plot(hole.x, hole.y, arduinoPin.x, arduinoPin.y)
+    .stroke({ width: 2, color: pinConnection.color, linecap: "round" });
+
+  line.data("component-id", componentId);
+  line.data("type", "wire");
+  line.data("update-wire", false);
+};
+
+export const createWireComponentToBreadboard = (
+  holeId: string,
+  componentEl: Element,
+  draw: Svg,
+  arduinoEl: Svg | Element,
+  componentConnectionId: string,
+  componentId: string,
+  color: string
+) => {
+  console.log(holeId, componentConnectionId, "look");
+  const hole = findBreadboardHoleXY(holeId, arduinoEl, draw);
+  const componentPin = findComponentConnection(
+    componentEl,
+    componentConnectionId
+  );
+  console.log(componentConnectionId, componentPin, "loop");
   const line = draw
     .line()
     .plot(hole.x, hole.y, componentPin.x, componentPin.y)
-    .stroke({ width: 2, color: color, linecap: "round" });
-  line.data("connection-id", connectionId);
-  line.data("component-id", element.id());
-  line.data("wire-type", type);
+    .stroke({ width: 2, color, linecap: "round" });
+  (window as any).WIRE_TEST = line;
+  line.data("connection-id", componentConnectionId);
+  line.data("component-id", componentId);
   line.data("type", "wire");
   line.data("update-wire", true);
-  line.data("hole-id", board.pinToBreadboardHole(pin));
-
-  return line;
+  line.data("hole-id", holeId);
 };
 
-export const createGroundWire = (
-  element: Element,
-  pin: ARDUINO_PINS,
-  arduino: Svg,
+export const createWireBreadboard = (
+  holeIdA: string,
+  holeIdB: string,
+  color: string,
   draw: Svg,
+  arduinoEl: Svg,
+  componentId: string
+) => {
+  const holeA = findBreadboardHoleXY(holeIdA, arduinoEl, draw);
+  const holeB = findBreadboardHoleXY(holeIdB, arduinoEl, draw);
+
+  const line = draw
+    .line()
+    .plot(holeA.x, holeA.y, holeB.x, holeB.y)
+    .stroke({ width: 2, color, linecap: "round" });
+
+  line.data("component-id", componentId);
+  line.data("type", "wire");
+  line.data("update-wire", false);
+};
+
+const getGroundorPowerWireLetter = (
+  isDown: boolean,
+  type: "ground" | "power"
+) => {
+  if (isDown && type === "power") {
+    return "W";
+  }
+
+  if (isDown && type === "ground") {
+    return "X";
+  }
+
+  if (!isDown && type === "power") {
+    return "Y";
+  }
+
+  return "Z";
+};
+
+export const createGroundOrPowerWire = (
+  hole: number,
+  isDown: boolean,
+  componentEl: Element,
+  draw: Svg,
+  arduino: Element,
   componentId: string,
-  direction: "left" | "right",
+  type: "ground" | "power"
+) => {
+  const groundHole = `pin${hole}${isDown ? "E" : "F"}`;
+  const pinConnectionId = type === "ground" ? "PIN_GND" : "PIN_POWER";
+  const color = type === "ground" ? "#000" : "#AA0000";
+  const breadBoardHoleA = `pin${hole}${isDown ? "A" : "J"}`;
+  const breadBoardHoleB = `pin${hole}${getGroundorPowerWireLetter(
+    isDown,
+    type
+  )}`;
+
+  createWireComponentToBreadboard(
+    groundHole,
+    componentEl,
+    draw,
+    arduino,
+    pinConnectionId,
+    componentId,
+    color
+  );
+
+  createWireBreadboard(
+    breadBoardHoleA,
+    breadBoardHoleB,
+    color,
+    draw,
+    arduino as Svg,
+    componentId
+  );
+};
+
+export const createComponentWire = (
+  hole: number,
+  isDown: boolean,
+  componentEl: Element,
+  pin: ARDUINO_PINS,
+  draw: Svg,
+  arduino: Element,
+  componentId: string,
+  connectionId: string,
   board: MicroController
 ) => {
-  return createBottomBreadboardWire(
-    element,
-    pin,
-    arduino,
+  const holeId = `pin${hole}${isDown ? "E" : "F"}`;
+  const breadboardHoleIdToBoard = `pin${hole}${isDown ? "A" : "J"}`;
+  console.log(board.pinConnections, pin, "record");
+  const color = board.pinConnections[pin].color;
+  createWireComponentToBreadboard(
+    holeId,
+    componentEl,
     draw,
-    "PIN_GND",
+    arduino,
+    connectionId,
     componentId,
-    "GND",
-    direction,
+    color
+  );
+
+  showPin(draw, pin);
+
+  createWireFromArduinoToBreadBoard(
+    pin,
+    arduino as Svg,
+    draw,
+    breadboardHoleIdToBoard,
+    componentId,
     board
   );
 };
 
-export const createPowerWire = (
-  element: Element,
-  pin: ARDUINO_PINS,
-  arduino: Svg,
-  draw: Svg,
-  componentId: string,
-  direction: "left" | "right",
-  board: MicroController
+export const findBreadboardHoleXY = (
+  pinHoleId,
+  arduino: Element,
+  draw: Svg
 ) => {
-  return createBottomBreadboardWire(
-    element,
-    pin,
-    arduino,
-    draw,
-    "PIN_POWER",
-    componentId,
-    "POWER",
-    direction,
-    board
-  );
+  const hole = findSvgElement(pinHoleId, draw);
+  return {
+    x: hole.cx() + arduino.x(),
+    y: hole.cy() + arduino.y(),
+  };
 };
 
-const createBottomBreadboardWire = (
-  element: Element,
-  pin: ARDUINO_PINS,
-  arduino: Svg,
-  draw: Svg,
-  componentBoxId: string,
-  componentId: string,
-  wireType: "POWER" | "GND",
-  direction: "left" | "right",
-  board: MicroController
+export const findResistorBreadboardHoleXY = (
+  holeId: string,
+  arduino: Element,
+  draw: Svg
 ) => {
-  const breadBoardLetter = wireType === "POWER" ? "W" : "X";
-  const wireColor = wireType === "POWER" ? WIRE_COLOR.POWER : WIRE_COLOR.GND;
-  const holeId = takeClosestBottomBreadboardHole(pin, direction, board);
-  const hole = findSvgElement(`pin${holeId}${breadBoardLetter}`, arduino);
-  const holeX = hole.cx() + arduino.x();
-  const holeY = hole.cy() + arduino.y();
+  const hole = findSvgElement(holeId, draw);
+  (window as any).hole = hole;
+  return {
+    x: hole.cx() + arduino.x(),
+    y: (hole.findOne("circle") as Element).cy() + arduino.y() - 1,
+  };
+};
 
-  const aHole = findSvgElement(`pin${holeId}A`, arduino);
+export const createResistorVertical = (
+  arduino: Svg | Element,
+  draw: Svg,
+  hole: number,
+  isDown: boolean,
+  componentId: string
+) => {
+  const resistorEl = draw.svg(resistorSvg).last();
+  resistorEl.data("component-id", componentId);
+  const holeId = `pin${hole}${isDown ? "D" : "I"}`;
 
-  const aHoleX = aHole.cx() + +arduino.x();
-  const aHoleY = aHole.cy() + +arduino.y();
+  const { x, y } = findResistorBreadboardHoleXY(holeId, arduino, draw);
+  resistorEl.cx(x);
+  resistorEl.y(y);
+};
 
-  const componentPin = findComponentConnection(element, componentBoxId);
-  (window as any).componentPin = componentPin;
-  const stationaryWire = draw
-    .line()
-    .plot(holeX, holeY, aHoleX, aHoleY)
-    .stroke({ width: 2, color: wireColor, linecap: "round" });
+export const resetBreadBoardHoles = (board: MicroController) => {
+  breadboard = _.cloneDeep(board.breadboard);
+};
 
-  stationaryWire.data("component-id", componentId);
-  stationaryWire.data("connection-id", componentBoxId);
-  stationaryWire.data("update-wire", false);
-  stationaryWire.data("hole-id", `pin${holeId}${breadBoardLetter}`);
-  stationaryWire.data("wire-type", wireType);
-  stationaryWire.data("type", "wire");
+export const showPin = (draw: Svg, pin: ARDUINO_PINS) => {
+  if (ANALOG_PINS.includes(pin)) {
+    const wire = draw.findOne(`#${pin}`);
+    if (wire) {
+      wire.show();
+    }
+  }
+};
 
-  const eHole = findSvgElement(`pin${holeId}E`, arduino);
-
-  const eHoleX = eHole.cx() + +arduino.x();
-  const eHoleY = eHole.cy() + +arduino.y();
-
-  const movingWire = draw
-    .line()
-    .plot(eHoleX, eHoleY, componentPin.x, componentPin.y)
-    .stroke({ width: 2, color: wireColor, linecap: "round" });
-
-  movingWire.data("component-id", componentId);
-  movingWire.data("connection-id", componentBoxId);
-  movingWire.data("update-wire", true);
-  movingWire.data("hole-id", `pin${holeId}E`);
-  movingWire.data("wire-type", wireType);
-  movingWire.data("type", "wire");
-
-  (window as any).line = stationaryWire;
-  return stationaryWire;
+export const hideAllAnalogWires = (draw: Svg) => {
+  ANALOG_PINS.forEach((pin) => {
+    const wire = draw.findOne(`#${pin}`);
+    if (wire) {
+      wire.hide();
+    }
+  });
 };
 
 export const updateWires = (element: Element, draw: Svg, arduino: Svg) => {
@@ -161,108 +279,4 @@ export const updateWires = (element: Element, draw: Svg, arduino: Svg) => {
       const componentPin = findComponentConnection(element, connectionId);
       w.plot(holeX, holeY, componentPin.x, componentPin.y);
     });
-};
-
-let bottomBreadBoardHoles: Array<{
-  status: "available" | "taken";
-  position: number;
-}> = [];
-
-export const takeNextBottomBreadboardHole = () => {
-  const hole = bottomBreadBoardHoles
-    .sort((holeA, holeB) => (holeA.position > holeB.position ? 1 : -1))
-    .find(({ status }) => status === "available");
-
-  hole.status = "taken";
-
-  return hole.position;
-};
-
-export const takeClosestBottomBreadboardHole = (
-  pin: ARDUINO_PINS,
-  direction: "right" | "left",
-  board: MicroController
-) => {
-  const pinNumber = parseInt(
-    board.pinToBreadboardHole(pin).replace("pin", "").replace("C", ""),
-    0
-  );
-
-  let sortedHoles = bottomBreadBoardHoles
-    .sort((a, b) => {
-      return (
-        Math.abs(a.position - pinNumber) - Math.abs(b.position - pinNumber)
-      );
-    })
-    .filter((hole) =>
-      direction === "right"
-        ? hole.position > pinNumber
-        : hole.position < pinNumber
-    );
-
-  if (sortedHoles.length === 0) {
-    sortedHoles = bottomBreadBoardHoles.sort((a, b) => {
-      return (
-        Math.abs(a.position - pinNumber) - Math.abs(b.position - pinNumber)
-      );
-    });
-
-    const selectedBreadBoardHole = sortedHoles.find(
-      (hole) => hole.status === "available"
-    );
-
-    selectedBreadBoardHole.status = "taken";
-    return selectedBreadBoardHole.position;
-  }
-
-  const selectedHole = sortedHoles.find((hole) => hole.status === "available");
-
-  selectedHole.status = "taken";
-  return selectedHole.position;
-};
-
-export const returnBottomHole = (position: number) => {
-  const index = bottomBreadBoardHoles.findIndex(
-    (hole) => hole.position === position
-  );
-
-  bottomBreadBoardHoles[index].status = "available";
-};
-
-export const resetBreadBoardHoles = (board: MicroController) => {
-  bottomBreadBoardHoles = _.range(4, 62)
-    .filter((i) => !board.skipHoles.includes(i))
-    .map((i) => {
-      return { status: "available", position: i };
-    });
-};
-
-export const findBreadboardHoleXY = (
-  pin: ARDUINO_PINS,
-  arduino: Element,
-  draw: Svg,
-  board: MicroController
-) => {
-  const hole = findSvgElement(board.pinToBreadboardHole(pin), draw);
-  return {
-    x: hole.cx() + arduino.x(),
-    y: hole.cy() + arduino.y(),
-  };
-};
-
-export const findResistorBreadboardHoleXY = (
-  pin: ARDUINO_PINS,
-  arduino: Element,
-  draw: Svg,
-  board: MicroController
-) => {
-  const hole = findSvgElement(
-    board.pinToBreadboardHole(pin).replace("E", "D").replace("F", "I"),
-    draw
-  );
-  (window as any).hole = hole;
-  return {
-    x: hole.cx() + arduino.x(),
-    y: (hole.findOne("circle") as Element).cy() + arduino.y() - 1,
-  };
 };
